requirements/development.txt
-r base.txt
django-debug-toolbar==4.3.0
django-extensions==3.2.3
ipython==8.22.1
